# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SCADA Logistics API",
    "description": "Мониторинг логистической инфраструктуры предприятия.\n\n**Авторизация:** `POST /auth/login` → `Authorization: Bearer <access_token>`; access живёт 15 мин, обновление — `POST /auth/refresh`. WebSocket: `/ws/live?token=<access_token>`. Демо-пользователи: `dispatcher`, `security`, `admin`, пароль `demo`.\n\nКоординаты — метры локального плана (см. `GET /objects`). Время — ISO 8601 с таймзоной.\n\nWebSocket `/ws/live`: формат сообщений — `contracts/ws.schema.json`.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/objects": {
      "get": {
        "tags": [
          "layout"
        ],
        "summary": "Get Objects",
        "description": "Вся статика одним ответом. Кэшируйте по `layout_version`: она растёт при каждом сохранении плана.",
        "operationId": "get_objects_objects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/layout": {
      "get": {
        "tags": [
          "layout"
        ],
        "summary": "Get Layout",
        "operationId": "get_layout_layout_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LayoutDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "layout"
        ],
        "summary": "Save Layout",
        "description": "Сохраняет новую версию плана и применяет её к реестру: здания, помещения, геозоны, дороги, КПП и\nположение неподвижных датчиков. Новые датчики на плане создаются с типовыми порогами; убранные с плана\nотключаются. Worker, коннекторы и симулятор подхватывают изменения в течение минуты.",
        "operationId": "save_layout_layout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LayoutUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LayoutDocument"
                }
              }
            }
          },
          "409": {
            "description": "The plan was saved by someone else since `base_version`"
          },
          "422": {
            "description": "Invalid plan; `detail` lists every problem"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Building": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "examples": [
              "b-wh2"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Склад №2 (холодный)"
            ]
          },
          "building_type": {
            "type": "string",
            "enum": [
              "office",
              "warehouse",
              "production",
              "garage"
            ],
            "title": "Building Type"
          },
          "floors": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Floors"
          },
          "geometry": {
            "$ref": "#/components/schemas/PolygonGeometry"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "building_type",
          "floors",
          "geometry"
        ],
        "title": "Building"
      },
...
      "Checkpoint": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "examples": [
              "cp-1"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "checkpoint_type": {
            "type": "string",
            "enum": [
              "vehicle",
              "pedestrian"
            ],
            "title": "Checkpoint Type"
          },
          "zone_id": {
            "type": "string",
            "title": "Zone Id"
          },
          "x": {
            "type": "number",
            "title": "X"
          },
          "y": {
            "type": "number",
            "title": "Y"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "checkpoint_type",
          "zone_id",
          "x",
          "y"
        ],
        "title": "Checkpoint"
      },
...
      "Geo": {
        "properties": {
          "x": {
            "type": "number",
            "title": "X"
          },
          "y": {
            "type": "number",
            "title": "Y"
          },
          "floor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Floor",
            "description": "Этаж внутри здания; null — улица"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "x",
          "y"
        ],
        "title": "Geo",
        "description": "Point in local plan coordinates, metres from the site origin (south-west corner)."
      },
      "Georef": {
        "properties": {
          "origin_lat": {
            "type": "number",
            "title": "Origin Lat"
          },
          "origin_lon": {
            "type": "number",
            "title": "Origin Lon"
          },
          "rotation_deg": {
            "type": "number",
            "title": "Rotation Deg",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "origin_lat",
          "origin_lon"
        ],
        "title": "Georef",
        "description": "Привязка локальных метров к WGS84: точка (0, 0) плана = (origin_lat, origin_lon)."
      },
      "Geozone": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "examples": [
              "z-gate"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "zone_type": {
            "type": "string",
            "enum": [
              "gate",
              "docks",
              "parking",
              "restricted",
              "speed"
            ],
            "title": "Zone Type"
          },
          "speed_limit_kmh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed Limit Kmh"
          },
          "geometry": {
            "$ref": "#/components/schemas/PolygonGeometry"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "zone_type",
          "geometry"
        ],
        "title": "Geozone"
      },
...
      "LayoutDocument": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "geojson": {
            "additionalProperties": true,
            "type": "object",
            "title": "Geojson",
            "description": "FeatureCollection в формате deploy/seed/layout.geojson"
          }
        },
        "type": "object",
        "required": [
          "version",
          "geojson"
        ],
        "title": "LayoutDocument"
      },
      "LayoutUpdate": {
        "properties": {
          "base_version": {
            "type": "integer",
            "title": "Base Version",
            "description": "Версия, от которой редактировали; при расхождении — 409"
          },
          "geojson": {
            "additionalProperties": true,
            "type": "object",
            "title": "Geojson"
          }
        },
        "type": "object",
        "required": [
          "base_version",
          "geojson"
        ],
        "title": "LayoutUpdate"
      },
      "LineGeometry": {
        "properties": {
          "type": {
            "type": "string",
            "const": "LineString",
            "title": "Type",
            "default": "LineString"
          },
          "coordinates": {
            "items": {
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "title": "Coordinates"
          }
        },
        "type": "object",
        "required": [
          "coordinates"
        ],
        "title": "LineGeometry"
      },
...
      "MetricSpec": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key",
            "description": "Ключ поля в payload события",
            "examples": [
              "temperature_c"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Температура"
            ]
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit",
            "examples": [
              "°C"
            ]
          },
          "kind": {
            "type": "string",
            "enum": [
              "number",
              "bool",
              "string"
            ],
            "title": "Kind",
            "default": "number"
          }
        },
        "type": "object",
        "required": [
          "key",
          "name"
        ],
        "title": "MetricSpec"
      },
      "ObjectsResponse": {
        "properties": {
          "layout_version": {
            "type": "integer",
            "title": "Layout Version"
          },
          "units": {
            "type": "string",
            "const": "m",
            "title": "Units",
            "default": "m"
          },
          "extent": {
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Extent",
            "description": "[x_min, y_min, x_max, y_max]"
          },
          "georef": {
            "$ref": "#/components/schemas/Georef"
          },
          "site": {
            "$ref": "#/components/schemas/PolygonGeometry"
          },
          "buildings": {
            "items": {
              "$ref": "#/components/schemas/Building"
            },
            "type": "array",
            "title": "Buildings"
          },
          "rooms": {
            "items": {
              "$ref": "#/components/schemas/Room"
            },
            "type": "array",
            "title": "Rooms"
          },
          "roads": {
            "items": {
              "$ref": "#/components/schemas/Road"
            },
            "type": "array",
            "title": "Roads"
          },
          "geozones": {
            "items": {
              "$ref": "#/components/schemas/Geozone"
            },
            "type": "array",
            "title": "Geozones"
          },
          "checkpoints": {
            "items": {
              "$ref": "#/components/schemas/Checkpoint"
            },
            "type": "array",
            "title": "Checkpoints"
          },
          "sensor_types": {
            "items": {
              "$ref": "#/components/schemas/SensorTypeInfo"
            },
            "type": "array",
            "title": "Sensor Types"
          },
          "sensors": {
            "items": {
              "$ref": "#/components/schemas/Sensor"
            },
            "type": "array",
            "title": "Sensors"
          },
          "vehicles": {
            "items": {
              "$ref": "#/components/schemas/Vehicle"
            },
            "type": "array",
            "title": "Vehicles"
          }
        },
        "type": "object",
        "required": [
          "layout_version",
          "extent",
          "georef",
          "site",
          "buildings",
          "rooms",
          "roads",
          "geozones",
          "checkpoints",
          "sensor_types",
          "sensors",
          "vehicles"
        ],
        "title": "ObjectsResponse",
        "description": "Вся статика одним ответом: план, реестр датчиков и машин. Фронт кэширует по layout_version."
      },
      "PolygonGeometry": {
        "properties": {
          "type": {
            "type": "string",
            "const": "Polygon",
            "title": "Type",
            "default": "Polygon"
          },
          "coordinates": {
            "items": {
              "items": {
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array",
                "maxItems": 2,
                "minItems": 2
              },
              "type": "array"
            },
            "type": "array",
            "title": "Coordinates",
            "description": "Кольца полигона в метрах плана; первое — внешнее"
          }
        },
        "type": "object",
        "required": [
          "coordinates"
        ],
        "title": "PolygonGeometry"
      },
...
      "Road": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "examples": [
              "road-main"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "width_m": {
            "type": "number",
            "title": "Width M"
          },
          "speed_limit_kmh": {
            "type": "number",
            "title": "Speed Limit Kmh"
          },
          "geometry": {
            "$ref": "#/components/schemas/LineGeometry"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "width_m",
          "speed_limit_kmh",
          "geometry"
        ],
        "title": "Road"
      },
...
      "Room": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "examples": [
              "r-wh2-storage"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "building_id": {
            "type": "string",
            "title": "Building Id"
          },
          "floor": {
            "type": "integer",
            "title": "Floor"
          },
          "room_type": {
            "type": "string",
            "title": "Room Type",
            "examples": [
              "storage"
            ]
          },
          "geometry": {
            "$ref": "#/components/schemas/PolygonGeometry"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "building_id",
          "floor",
          "room_type",
          "geometry"
        ],
        "title": "Room"
      },
...
      "Sensor": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z0-9][a-z0-9_-]*$",
            "title": "Id",
            "examples": [
              "clim-wh2-storage"
            ]
          },
          "type": {
            "$ref": "#/components/schemas/SensorType"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name",
            "examples": [
              "Климат, склад №2: хранение"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "building_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Building Id",
            "examples": [
              "b-wh2"
            ]
          },
          "zone_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone Id",
            "examples": [
              "r-wh2-storage"
            ]
          },
          "floor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Floor"
          },
          "geo": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Geo"
              },
              {
                "type": "null"
              }
            ],
            "description": "Точка установки; null для подвижных датчиков"
          },
          "vehicle_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicle Id",
            "description": "Для GNSS-трекера: машина, на которой он стоит"
          },
          "thresholds": {
            "items": {
              "$ref": "#/components/schemas/Threshold"
            },
            "type": "array",
            "title": "Thresholds",
            "default": []
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "is_mobile": {
            "type": "boolean",
            "title": "Is Mobile"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "name",
          "is_mobile",
          "created_at"
        ],
        "title": "Sensor"
      },
...
      "SensorType": {
        "type": "string",
        "enum": [
          "anpr_camera",
          "access_control",
          "gnss",
          "motion",
          "climate"
        ],
        "title": "SensorType"
      },
      "SensorTypeInfo": {
        "properties": {
          "id": {
            "$ref": "#/components/schemas/SensorType"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Климат (температура и влажность)"
            ]
          },
          "is_mobile": {
            "type": "boolean",
            "title": "Is Mobile",
            "description": "Подвижный датчик: позиция приходит в событиях, а не задана на плане"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/MetricSpec"
            },
            "type": "array",
            "title": "Metrics"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "is_mobile",
          "metrics"
        ],
        "title": "SensorTypeInfo"
      },
...
      "Threshold": {
        "properties": {
          "metric": {
            "type": "string",
            "title": "Metric",
            "examples": [
              "temperature_c"
            ]
          },
          "nominal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nominal",
            "description": "Номинальное значение",
            "examples": [
              4.0
            ]
          },
          "min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min",
            "examples": [
              2.0
            ]
          },
          "max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max",
            "examples": [
              6.0
            ]
          },
          "critical_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Critical Min",
            "examples": [
              0.0
            ]
          },
          "critical_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Critical Max",
            "examples": [
              8.0
            ]
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version",
            "description": "Версия порога; алерт хранит версию, по которой сработал",
            "default": 1
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "metric"
        ],
        "title": "Threshold",
        "description": "Пороги одной метрики. Норма — [min, max]; вне нормы — warning; вне [critical_min, critical_max] — critical."
      },
...
      "Vehicle": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "examples": [
              "v-truck-1"
            ]
          },
          "plate": {
            "type": "string",
            "title": "Plate",
            "examples": [
              "А123ВС77"
            ]
          },
          "kind": {
            "$ref": "#/components/schemas/VehicleKind"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "examples": [
              "КАМАЗ 65115"
            ]
          },
          "sensor_id": {
            "type": "string",
            "title": "Sensor Id",
            "description": "GNSS-трекер машины",
            "examples": [
              "gnss-truck-1"
            ]
          },
          "carrier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carrier",
            "examples": [
              "ООО «ТрансЛогистик»"
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "plate",
          "kind",
          "model",
          "sensor_id"
        ],
        "title": "Vehicle"
      },
      "VehicleKind": {
        "type": "string",
        "enum": [
          "truck",
          "loader",
          "car"
        ],
        "title": "VehicleKind"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}